#706 Refactor Transpile class. Separate transpile, validation and generation of Java files
